+++ aquacrop/native/inflow_w.f90
@@ -23,7 +23,7 @@ contains
     real(real64), dimension(n_cell), intent(in) :: z_gw
     real(real64), dimension(n_cell, n_layer, n_crop, n_farm), intent(in) :: th_sat
     real(real64), dimension(n_comp), intent(in) :: dz
-    real(real64), dimension(n_comp), intent(in) :: layer_ix
+    integer(int32), dimension(n_comp), intent(in) :: layer_ix
     integer(int32) :: i, j, k
     do i = 1, n_farm
        do j = 1, n_crop
