+++ aquacrop/native/rainfall_partition.f90
@@ -40,7 +40,7 @@ contains
     real(real64), intent(in) :: cn_top
     real(real64), dimension(:), intent(in) :: dz
     real(real64), dimension(:), intent(in) :: dz_sum
-    real(real64), dimension(:), intent(in) :: layer_ix
+    integer(int32), dimension(:), intent(in) :: layer_ix
 
     integer(int32) :: max_comp_idx
     integer(int32) :: i
